Organizacion de modelos y creacion de Icons.
