+++ lib/components/detail_expense.dart
@@ -1,3 +1,4 @@
+import 'package:app_wallet/models/category.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:app_wallet/models/expense.dart';
