+++ lib/components/widgets/chart/chart.dart
@@ -1,4 +1,6 @@
 import 'package:app_wallet/library/main_library.dart';
+import 'package:app_wallet/models/category.dart';
+import 'package:app_wallet/models/expense_bucket.dart';
 
 class Chart extends StatelessWidget {
   const Chart({super.key, required this.expenses});
