+++ lib/components/widgets/expenses_list/expense_item.dart
@@ -1,3 +1,4 @@
+import 'package:app_wallet/models/category.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart'; // Importa el paquete intl
 import 'package:app_wallet/models/expense.dart';
