+++ lib/models/category.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+enum Category { comida, viajes, ocio, trabajo, salud , servicios }
+
+const categoryIcons = {
+  Category.comida: Icons.lunch_dining,
+  Category.viajes: Icons.flight_takeoff,
+  Category.ocio: Icons.movie,
+  Category.trabajo: Icons.work,
+  Category.salud: Icons.health_and_safety,
+  Category.servicios: Icons.design_services,
+};
