+++ lib/models/expense.dart
@@ -1,21 +1,10 @@
-import 'package:flutter/material.dart';
-import 'package:uuid/uuid.dart';
+
 import 'package:intl/intl.dart';
+import 'package:uuid/uuid.dart';
+import 'package:app_wallet/models/category.dart';
 
 final formatter = DateFormat.yMd();
-
 const uuid = Uuid();
-
-enum Category { comida, viajes, ocio, trabajo, salud , servicios}
-
-const categoryIcons = {
-  Category.comida: Icons.lunch_dining,
-  Category.viajes: Icons.flight_takeoff,
-  Category.ocio: Icons.movie,
-  Category.trabajo: Icons.work,
-  Category.salud: Icons.health_and_safety,
-  Category.servicios: Icons.design_services,
-};
 
 class Expense {
   Expense({
@@ -35,28 +24,4 @@ class Expense {
     return formatter.format(date);
   }
 }
-
-class ExpenseBucket {
-  const ExpenseBucket({
-    required this.category,
-    required this.expenses,
-  });
-
-  ExpenseBucket.forCategory(List<Expense> allExpenses, this.category)
-      : expenses = allExpenses
-            .where((expense) => expense.category == category)
-            .toList();
-
-  final Category category;
-  final List<Expense> expenses;
-
-  double get totalExpenses {
-    double sum = 0;
-
-    for (final expense in expenses) {
-      sum += expense.amount; // sum = sum + expense.amount
-    }
-
-    return sum;
-  }
-}
+
